Update 주유소.java
58점짜리 정답안을 100점짜리 정답안으로 수정

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -4,8 +4,8 @@ import java.io.*;
 public class Main {
     
     public String solution(int n, int[] dis, int[] prices) {
-        int answer = 0;
-        int cntMinPrice = prices[0];
+        long answer = 0L;
+        long cntMinPrice = prices[0];
         for(int i=0; i<dis.length; i++) {
             if(cntMinPrice > prices[i]) {
                 cntMinPrice = prices[i];
